Fix warnings due to improved CCOPT_Linux in c.mk.

+++ t_malloc/killer.c
@@ -4,10 +4,14 @@
 
 int main (int argc, char *argv[]) {
 
-  int pid=atoi(argv[1]);
+  int pid = atoi(argv[1]);
 
-  for (;;) {
-    kill (pid, SIGUSR1);
+  if (argc == 2) {
+    pid = atoi(argv[1]);
+
+    for (;;) {
+      kill (pid, SIGUSR1);
+    }
   }
   exit(0);
 }
